Handle slow architecture failures
After trying to wait for hidden elements and other wait strategies
I found this useful.

Popup 'Before starting the installation ...' needs to be taken into
account when it disappears as it changes position of other elements
in the screen.

Storage in Overview section also appears later and changes position
of other elements.