Handle Overview after registration
After registration Agama goes to configuration screen and later to
Overview, but sometimes with Puppeteer we go too fast through the
following steps after registration and the page is ready to process
those steps, like creation of user, but the signal to go to
configuration/overview page arrives later breaking the automation.

At the moment we only check that the alert is not there, but that
is not enough, we need to ensure to be in Overview.

Agama developer plan to polish this peculiar behaviour in 16.1
but take into account that it is only something that happen with
very fast automation, a user rearly would encounter this.